11- Doing Tests, added Mapper into project.

+++ src/main/java/com/borrow/borrowsecurity/product/ProductController.java
@@ -2,7 +2,6 @@ package com.borrow.borrowsecurity.product;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -18,7 +17,7 @@ public class ProductController {
     private final ProductService productService;
 
     @GetMapping
-    public ResponseEntity<Product> getProduct(@RequestParam Integer id){
+    public ResponseEntity<Product> getProduct(@RequestParam Long id){
         return ResponseEntity.ok(productService.getProduct(id));
     }
